Create KNN classifier
KNN stands for K nearest neighbor. It's a supervised learning algorithm. In this example, I use it as a classifier to identify 3 classes. 
* mask worn correctly
* mask under nose
* no mask

You can use this code and edit it on the P5JS editor but you'll need to train it for a few minutes (or seconds). The classes can be changed. it's not actually built to detect masks but to classify anything. You can use it to classify cancerous and non-cancerous mole (good luck finding a decent dataset for free) or bananas from apples. You can also us it to differentiate different foods like "hotdog" and "not hotdog".