Create review analyzer
As a product manager, it is crucial to know what people think about your product. However, reading every review can take a lot of time and most of the tools that offer that feature, such as MonkeyLearn, are very expensive. We've built a tool that helps you get a quick glimpse of what your users talk about when they mention your product. The Sentiment feature helps determin the tone of their review. The categories help you know the general topic of the review. Finally, the keyword detection feature lets you know which keywords are most common.